requirements.txt
asgiref==3.5.2
atomicwrites==1.4.1
attrs==21.4.0
colorama==0.4.5
coverage==6.4.2
Django==4.0.6
djangorestframework==3.13.1
Faker==12.0.1
iniconfig==1.1.1
mixer==7.2.2
packaging==21.3
pluggy==1.0.0
psycopg2==2.9.3
py==1.11.0
PyJWT==2.4.0
pyparsing==3.0.9
pytest==7.1.2
pytest-cov==3.0.0
pytest-django==4.5.2
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
tzdata==2022.1
